+++ test/dummy/config/application.rb
@@ -4,7 +4,6 @@ require "active_model/railtie"
 require "active_record/railtie"
 require "action_controller/railtie"
 require "action_view/railtie"
-require "protected_attributes" # Rails 4 requirement for using `attr_protected` and `attr_accessible`
 
 Bundler.require(:default, Rails.env) if defined?(Bundler)
 require 'paper_trail'
@@ -51,7 +50,7 @@ module Dummy
     # This will create an empty whitelist of attributes available for mass-assignment for all models
     # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
     # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = false
+    config.active_record.whitelist_attributes = false if ::PaperTrail.active_record_protected_attributes?
 
     # Enable the asset pipeline
     config.assets.enabled = false