+++ test/test_helper.rb
@@ -3,14 +3,13 @@ ENV["RAILS_ENV"] = "test"
 
 require File.expand_path("../dummy/config/environment.rb",  __FILE__)
 require "rails/test_help"
-require "minitest/reporters"
-MiniTest::Reporters.use! MiniTest::Reporters::SpecReporter.new
 #ActionMailer::Base.delivery_method = :test
 #ActionMailer::Base.perform_deliveries = true
 #ActionMailer::Base.default_url_options[:host] = "test.com"
 
 Rails.backtrace_cleaner.remove_silencers!
 
+require 'shoulda'
 require 'ffaker'
 
 # Run any available migration